+++ lib/src/televerse/context/context.dart
@@ -44,7 +44,7 @@
 /// - [ChatMemberUpdatedContext] - Represents the context when a [UpdateType.chatMember] or [UpdateType.myChatMember] event occurs.
 /// - [PollContext] - Represents the context when a [UpdateType.poll] event occurs.
 /// - [PollAnswerContext] - Represents the context when a [UpdateType.pollAnswer] event occurs.
-///
+/// - [ChosenInlineResultContext] - Represents the context when a [UpdateType.chosenInlineResult] event occurs.
 /// More context types will be added in the future.
 library televerse.context;
 
@@ -61,6 +61,7 @@ part 'callback_query.dart';
 part 'chat_member.dart';
 part 'poll.dart';
 part 'poll_answer.dart';
+part 'chosen_inline_result.dart';
 
 /// This class is used to represent the context of an update. It contains the update and the [RawAPI] instance.
 ///
@@ -72,6 +73,7 @@ part 'poll_answer.dart';
 /// - [ChatMemberUpdatedContext] - Represents the context when a [UpdateType.chatMember] or [UpdateType.myChatMember] event occurs.
 /// - [PollContext] - Represents the context when a [UpdateType.poll] event occurs.
 /// - [PollAnswerContext] - Represents the context when a [UpdateType.pollAnswer] event occurs.
+/// - [ChosenInlineResultContext] - Represents the context when a [UpdateType.chosenInlineResult] event occurs.
 ///
 /// Contexts are subclasses of this class. You can use this class to access the update and the Televerse instance.
 class Context {