+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added `Televerse.pollAnswer` method to listen to `poll_answer` updates.
 - Now most of the Televerse methods return a `StreamSubscription` instead of `void`.
 - This is to make it easier to cancel the subscription and increase the performance.
+- Added `ChosenInlineResultContext` class to represent the context of a `chosen_inline_result` update and `Televerse.chosenInlineResult` method to listen to `chosen_inline_result` updates.
 ## 1.8.1
 - Added `ChatMemberUpdatedContext` class to represent the context of a `chat_member` and `my_chat_member` updates.
 - Added `Televerse.chatMember` and `Televerse.myChatMember` methods to listen to `chat_member` and `my_chat_member` updates.